Title: 실수 받아 그대로 출력, Time: 144ms, Memory: 11MB, Status: Passed - Codetree

+++ 260128/실수 받아 그대로 출력/enter-real-value-and-print.java
@@ -5,7 +5,7 @@ public class Main {
         // Please write your code here.
         Scanner sc=new Scanner(System.in);
 
-        double N=sc.nextInt();
+        double N=sc.nextDouble();
 
         System.out.printf("%.2f",N);
     }
